fix: fix loading of env file in packaged app (#189)
Towards #97 

We have to tell dotenv the env file's location when it's not in the same
directory as the process that starts the app.

Also uses this opportunity to simplify the build process a bit. No
longer need a distinct `.env.production` file, as the plan is to
generate a `.env` file in a github workflow step prior to building the
app.