chore: introduce improvements to the forge configuration for building and packaging (#198)
Towards #101 

Most of the changes are related to getting relevant information about
the kind of app that's running to the main process, which will be used
to configure Sentry appropriately. As I worked on this, I dealt with a
variety of rather annoying inconsistencies related to how Forge
configures an application's name and how Electron does (not) use it. To
summarize, this PR:

- introduces app.config.json fields called `appType` and `appVersion`.
`appType` indicates to the kind of application that's running e.g. dev,
internal, release candidate, production. `appVersion` indicates the
user-facing version of the application that's running. This is mostly
meant to be used for having a version string that is aligned with our
custom versioning format (for releases) and provides useful debugging
info when desirable.

- introduces more explicit handling of the app name that's configured by
forge and that's used by electron internally. through lots of
trial-and-error, I found that configuring the app name using
`@electron/packager` via forge does not actually affect the what's used
by Electron at runtime. It's mostly used for things at build time (e.g.
build executable name, asset names, pList info). Frustratingly, Electron
doesn't use what's configured in packager and instead uses the values
found in the `package.json` file, which unfortunately is not desirable
due to the dynamic nature of app names we'd like to support. In addition
to that, at least on macOS, the name of the application is actually
incredibly important when it comes to how data isolation is handled.
Without any changes, I realized that I kept running into decryption
issues like `Error while decrypting the ciphertext provided to
safeStorage.decryptString` because:

1. The app name was always `comapeo-desktop`, thus reading from the same
user data directory regardless of the application type.
2. Electron stores keychain entries based on the app name. Due to (1),
it would store an entry called `comapeo-desktop Safe Storage` regardless
of the app type, which eventually leads to issues with decryption if
you're switching between different app types.
  
With the changes, we now create applications that have distinct names
based on the app type. For example, in development we'll have an app
called `CoMapeo Desktop Dev` whereas a release candidate will be called
`CoMapeo Desktop RC`. At least on macOS, due to the different naming
these applications have completely separate application directories,
which is the desired goal.