chore: define `__APP_TYPE__` variable for renderer app to use (#199)
Towards #101 

This basically makes the `APP_TYPE` env variable available for the
renderer app, which will eventually be used to configure Sentry
appropriately.

The reason for doing it as a defines instead of a `VITE_...` env
variable is to ensure that the source of this variable comes from the
`APP_ENV` variable, which in practice is provided to the renderer
through the forge config that calls the build process for the vite app.