chore: fix a variety of issues with the create-build workflow (#212)
Was trying to test out manually triggering this on one of my existing
PRs and ran into a myriad of issues.

- Fixes the uploaded artifact name. Before it used to just be
`<platform>-` i.e. missing the SHA. Can do a follow-up to use a
shortened SHA instead but for now, fine to use the full one.

- Fixes an issue where `forge make` fails in CI due to us throwing a
top-level error in the forge config when an env file isn't present.

- Fixes an issue where we don't upload the debian package artifact.

- Updates the workflow to not fail fast. Much easier to get a more
productive feedback loop with this.

- Fixes errors building on linux due to the existence of native
prebuilds that are incompatible with the platform

- Fixes an issue on windows where the node script to generate the
translated languages json file fails.

- Fixes an issue on windows where renaming a directory to another
directory fails.

- Fixes an issue on windows where the squirrel maker complains about not
being able to use the image configured for setting up an icon. Need to
create a follow-up issue to update the asset since it has to be ✨
special ✨ and can't just be a renamed png file (which is what i did lol)

- Fixes an issue on linux where the executable name to use for packaging
is invalid.

Remaining issues:

- Seems like the intel macos build sometimes fails. See
https://github.com/create-dmg/create-dmg/issues/190
- The created distributable name doesn't include the build SHA , which
is preferable to have for the internal and release candidate builds.
Will address as a separate task since it's probably a little tricky to
fix.