chore: improve Sentry configuration (#218)
Follow-up to #101

Based on
https://github.com/digidem/comapeo-desktop/pull/200#pullrequestreview-3120819981,
which came in after merging the PR.

The only suggestion I didn't implement was the usage of a npm module to
do the process arg parsing in the preload, as I realized that we cannot
access npm-installed modules in the preload scripts, contrary to what I
thought I read in the docs. If we disable the sandboxing for the window
or bundle the preload script then we can achieve this, but I don't think
it's useful to do that work right now.