feat: allow creation of track in test data creation page (#224)
Towards #161

Adds a toggle in the test creation page to enable creating a track that
uses the observations created as its reference. We don't try to do
anything super smart in terms of generating realistic location points
for the track, and instead just use the observations' coordinates as the
track's recorded locations.