fix: fix executable name for macOS (#243)
By convention, the executable name for the application on macOS is
usually the same as the application name. However, I thought it would be
okay for it to adhere to more conventional posix-naming i.e. lowercase +
hyphen-separated. While this generally works, some applications will
display the executable name instead of the application name, such as
Raycast. On Finder, seems like it'll default to the name of the
application directory found in the `Applications` folder, so it's not an
issue for that at least. However, to potentially avoid similar cases
with other tools, I think it's safer to just follow the convention
instead of being more posix-friendly on macOS 😄