fix: fix app resizing issue when going into fullscreen mode (#245)
~Not sure why, but the relevant browser environment variable seems to go
missing when the app enters fullscreen mode. Maybe some undocumented
behavior or an issue in Electron?~ EDIT: think it's a case of me not
reading the - admittedly scattered - docs
[here](https://github.com/WICG/window-controls-overlay/blob/main/explainer.md#stylecss):

> The height is set to titlebar-area-height or to fall back to
var(--fallback-title-bar-height) if the window controls overlay is not
visible.

Guess that variable isn't present if the controls go missing, as they do
when entering fullscreen mode on macOS.