chore(deps): update deps (#256)
Most notably updates Electron to 37.4.0, which means the Node version
has been updated to 22.18.0 (via
[37.3.0](https://github.com/electron/electron/releases/tag/v37.3.0)).
This means that we'll be able to use Node's built-in type stripping for
build tooling and main process code (see
https://nodejs.org/en/blog/release/v22.18.0)! I experimented with this
locally and the results are very promising. Will revisit it and create a
separate PR for it if I feel it's good enough to use.

Unfortunately, moving to 22.18.0 for the dev environment introduces a
notable performance regression in npm script execution, especially for
some of our more involved scripts. I've reported an issue to npm (see
https://github.com/npm/cli/issues/8533). In the meantime, I have a
separate piece of work that's done that allows us to make use of node's
built-in script runner to mitigate some of the regressions. Will create
a PR shortly.

Another notable upgrade is ESLint to 9.34.0, which introduces
multithread linting (see
https://eslint.org/blog/2025/08/multithread-linting/). I gave it a try
but found that it didn't bring any notable benefits for us. In fact, it
was slower to specify a concurrency greater than 1. Good to know about
for future reference though!