chore: use zustand for persisted main process store implementation (#260)
_NOTE: this may be a terrible idea. still debating..._

Basically replaces our usage of `electron-store` with `zustand` and its
persist middleware. There were a few issues I had with `electron-store`:

- It provides functionality that I don't envision us ever needing e.g.
renderer process integration, watching state changes across processes,
encryption, etc.

- It has its own data validation logic that's based on ajv internally.
Getting that to play nicely with the validation we do with valibot
without things getting redundant or verbose was getting increasingly
difficult. It's also much more limited from a type-safety perspective
than this new approach that takes advantage of valibot quite
effectively.

- Based on issues I saw in their repo, their migration functionality
seems a bit shaky. We didn't even have anything in place with it, as I
wasn't confident in using it.

- It blocks #259, which provides greater benefit than this module
currently brings us.

Why Zustand? Mostly due to familiarity, a small footprint, and
simplicity of its API. It does feel a bit odd to use this in the context
of a NodeJS process, but I guess that speaks to its flexibility. I did
consider doing an implementation that was a bit more barebones, but then
realized that implementing all of the features related to persistence
would be a massive chore that Zustand and its persistence middleware
already handle. I'm not too worried about any potential performance
issues with this approach compared to something that may use native
primitives (e.g. event emitter) since this is not being called in
hotpaths or overly expensive to use.

It's important to note that this switch will cause issues for those who
have used the app in the wild (whether installed officially or used via
dev environment), as the content schema of the file has been updated to
account for schema versioning (to enable seemless migrations in the
future). Basically goes from just the state to something like:

```json
{
  "version": 1,
  "state": ...
}
```

I'm not worried about this issue as the app is not officially released.
The recommended course of action is to uninstall the app or clear the
data directory and start anew. We would have to do this at some point
before official release anyways, so it feels appropriate to do it here.

Other than that, there should be no noticeable changes in app behavior!