chore: update and pin GitHub actions (#262)
Pinning is a precautionary measure against the security loophole of
using version tags for actions (tags are mutable and thus can be pointed
at bad commits if there's a compromise). More details here:
https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions

Used a combination of https://github.com/suzuki-shunsuke/pinact and
https://github.com/azat-io/actions-up to do this. In most cases, can
probably rely on the latter for doing upgrades.