feat: add some basic map controls and elements (#263)
Closes #239 

- Adds a scale bar (bottom left)
- Adds zoom in/out controls (top right)
- Adds "zoom-to-data" control (top right). Only available when not
viewing observation/track details.

Didn't add the compass control for now. Didn't think it was necessary
given that we cannot rotate the map (yet), as pressing it wouldn't do
anything. Could be mistaken about whether it's still valuable to include
it. Can add as a follow-up if necessary.

Can do more customization of the actual controls if desired, but seems
like we're okay with MapLibre's default styles for now.