fix: redirect user to an existing project when active project ID is missing (#271)
If the `activeProjectId` is missing on app start, we typically redirect
the user to the onboarding flow, which is >99% of the cases that will
happen. There's a <1% chance that the user has already created/joined a
project but somehow the `activeProjectId` is missing, in which case it's
preferable to prevent the user from going through the onboarding flow.
This PR addresses this by redirecting the user to _some_ valid project.
It's an edge case and the solution here feels like a short-term one. In
the long-term, I think it'd be better to redirect the user to some
project selection page where they can decide which project to enter (or
even create/join a new one)