chore: use isolated directory for storing comapeo data (#279)
We now specify a dedicated directory within the user data directory for
the backend to use instead of the root of the user data directory. this
is for future proofing as we don't have control over what electron
stores in there and how it names things.

note that this means that end users who have installed and run the app
will be affected, basically as though they are a new user. the
recommendation is to clear the relevant directory or completely
uninstall the application. Not worried about this as we are not doing
official releases. better to have this done now.