chore: update windows packaging configuration (#280)
Came across this detail in the [Forge
docs](https://www.electronforge.io/config/makers/squirrel.windows#spaces-in-the-app-name)
and it became a small rabbithole of updating the app configuration to
account for it.

> Squirrel.Windows can behave unexpectedly when application names
contain spaces.

There's a change to the app config schema as a result but didn't treat
it as something worth writing a migration for, given that it's optional,
only applies to windows, and we're still in the pre-release phase.