chore: add improvements to intl setup in main process (#294)
Adding localization support for things managed in the main process is
still in its early stages and not used meaningfully just yet. Was
revisiting it as I'm thinking about translations for native controls
like the app menu and context menu, and noticed a couple of issues and
shortcomings.

This PR updates our setup in the following ways:

- renames our custom `Intl` class to `IntlManager`. `Intl` is a global
namespace object so we rename our implementation to avoid confusion.

- Make the `IntlManager` an event emitter. This will be useful for
regenerating the user-facing translated labels of native UI. We emit a
`locale-state` event that provides the locale state as the event payload
when it changes.

- fixes the loading of translations. before we were using JSON module
import syntax and only loading the messages for a hardcoded list of
english, spanish, and portuguese. now we attempt to load and cache the
messages based on the language of interest, with english being the only
one that we load ahead of time.