feat: add basic context menu (#295)
When you right-click on the screen, a context menu with some common and
useful actions appears. The following items are available:

- `Copy`
- `Copy Image`
- `Copy Image Address`
- `Copy Link`
- `Cut`
- `Inspect Element` (only enabled in non-production builds)
- `Learn Spelling`
- `Look up`
- `Paste`
- `Save Image As`
- `Select All` (only on Windows and Linux)

We may need to adjust this depending on UX requirements, which I can do
as follow-ups whenever desired.

The items that appear vary based on what you right-click on. The
underlying library we introduced for this handles most of the common
cases when it comes to text selection, media, etc.