feat: implement app diagnostics metrics (#300)
Towards #164 

- Introduces a scheduling abstraction called `MetricsScheduler` that's
used to encompass the shareable scheduling logic implemented in mobile.
The intention is to eventually extract this out to a separate library to
use between the applications. The current scope of this is only for
metrics that have the "send at most once per day" requirement, which is
relevant for app diagnostics and project stats. Note that this
abstraction is implemented such that it follows the storage patterns
that the mobile implementation uses so that the eventual library can be
drop-in for mobile.

- Unlike the mobile implementation, the implementation for app
diagnostics here does not account for the following:
- app state changes: this is not really relevant for desktop
applications
- changes in the network: the mobile implementation keeps some internal
state about internet access and tries to adjust based on that, which I
think is not necessary for desktop right now. I made the decision for
network connectivity to be a concern of the scheduler abstraction's
`send` parameter if needed.

- Since the mobile implementation of device diagnostics has a different
schedule and implementation, going to do a follow-up PR that basically
does a 1-to-1 port from mobile.

- I updated the repo secrets to specify
`COMAPEO_DIAGNOSTICS_METRICS_URL` and `COMAPEO_METRICS_ACCESS_TOKEN`.