feat: add collaborator info page and ability to leave project (#305)
Closes #203 

Also includes relevant updates to the UI in the team page.

As noted in the issue, there's a known issue where attempting to leave
the project as the last coordinator does not work. This needs to be
fixed in core though so no adjustments to the implementation were made
to account for this.