fix: fix navigation issues related to post-onboarding invite flow (#312)
Fixes #248 

Basically needed to meticulously make sure that navigations within the
flow are replacing the URL as opposed to pushing new entries. Kind of
tedious and annoying but it's a much smaller diff than I was expecting.

Also fixes an issue with some strings not being set up for translation.