chore: fix url for build-rc comment (#319)
Basically undoes #316 as I forgot that the artifact URL output is per
matrix option, which doesn't work for our needs. Also causes a warning
from GitHub that probably would've caused things to not work anyways:

<img width="502" height="344" alt="image"
src="https://github.com/user-attachments/assets/7d1960b2-1e2b-46e2-a342-66341f649006"
/>

Instead we opt for getting the action run URL and pointing to that.
Updates the comment a bit to make it clearer where in that page to look.