chore: specify dev engines field in package.json (#321)
Makes use of the
[`devEngines`](https://docs.npmjs.com/cli/v11/configuring-npm/package-json#devengines)
field which allows specifications for development tooling which has the
following purposes:

- consistent development environments for contributors and workflows
- standardized information that tools can use (e.g. actions, other cli
tools, etc).

Ideally we'd set things to error when conditions are not met, but I do
worry that it may be a bit too restrictive to do that for now (also
won't play nicely with CI workflows right now).

Ideally `setup-node` would support this so we don't need to update the
relevant workflows with an extra step but it doesn't yet (see
https://github.com/actions/setup-node/issues/1255). It's not too much of
a lift to do this manually for now.