fix: update text in network connection info component (#344)
Closes #330

The decided solution is more of a workaround as opposed to a fix to the
underlying issue. The module we use to get access to the connected wi-fi
network is prone to flakiness so instead of trying to solve it at that
layer, we opt for updating the copy to make the failure situation less
misleading.