fix: fix various navigation issues (#355)
There were a couple of issues found in QA related to navigation:

1. Leaving a project results in getting stuck in a non-functioning page.

This was mostly caused by a combination of our usage of `reloadDocument:
true` with `router.navigate()` in the built app. The navigate method
does not account for the hash history when it comes to constructing the
final URL to navigate to, resulting in a URL that would be missing the
actual file in the path name. For example, instead of a navigating to
`file://Some/Path/index.html#/onboarding/project`, we were trying to
navigate to `file:///onboarding/project`. This becomes problematic with
`reloadDocument: true` as it tries to do a full page reload, which fails
because that it points to a file that doesn't exist. The fix was to
adjust our usage of `router.navigate()` such that we're building the
correct URL that accounts for the hash history router.

2. Attempting to leave a project fails after leaving one prior.

This was a gnarly one. Basically, there's a known issue in Tanstack
Router related to path params not being correctly matched to non-nested
routes with path params. In our case, we have a situation where we have
routes at `/app/projects/$projectId/...` but also
`/app/projects/$projectId_/...`. The route for leaving a project lives
under the latter and we were running into a case where the `$projectId`
param was stale. The following steps resulted in being stuck in a
infinite loop:

    1. Create project 1
    2. Leave project 1
    3. Create project 2
    4. Leave project 2
    
In step 4, what was happening is that we were navigating to project 1
instead of project 2 and attempting to leave project 1 again, which
causes a failure that we don't handle well. Might be worth looking into
more elegantly handling this either here or in core, but this shouldn't
be possible to attempt to begin with. The fix for this is enabling an
experimental option in the router plugin that fixes this exact issue.
It's only experimental because it results in notable breaking changes
that require work for larger apps, but for us it's not an issue. The
plan is for it to be enabled by default in the next major version
anyways.
    
In addition to the aforementioned fixes, this PR also updates the state
model for the "active project ID". Before, we were representing this as
async state with tanstack query since it's persisted in the backend.
However, I ran into invalidation issues in the router context that I
believe are related to it being async. Instead, we update it to be
synchronous state (using zustand) that we persist to the backend as an
effect. The persisted value is only used directly by the app as an
initial zustand store state value through browser window init flag. This
should make it more predictable to work with in router lifecycles moving
forward, as we now provide the store to the router context as opposed to
just the reactive store state.