fix: fix getting initial project ID when doing manual browser page reload (#356)
Follow-up to #355 . Fixes an issue that's only relevant when using the
app for the first time. Can reproduce by doing the following:

1. Open app for first time.
2. Go through onboarding
3. Enter app
4. Manually refresh page e.g. `ctrl + r`

What's happening is that the whole app reloads but the initial window
arguments provided when creating the browser window are not
recalculated. So after going through the onboarding and having a valid
active project ID that's persisted, if you reload the page, we
initialized the store value using the stale value that comes from
opening the app fresh for the first time (undefined). This issue is not
relevant if you close the app after step 3, re-open it, and perform step
4.

Solution is a bit ugly but we provide a fallback IPC call to
synchronously get the stored active project ID if the initial window
argument is undefined. This allows us to keep the same setup code
without potentially exposing more runtime methods that get intermixed in
the app loaders.