fix: fix more issues related to loading active project ID (#368)
Follow-up to #355 

Fixes the following issues:

### Leaving a project causes navigation to stale/invalid project

Was happening under the following sequence:
  
1. Go through onboarding flow (join or create), but before navigating to
the "main" app, close the application.
2. Reopen navigation
3. Leave project 

Issue was that the active project id was not being persisted before
closing the app in this case. When re-opening the app, the initial
project id is calculated on page load based on the existing projects,
since nothing's been persisted. this results in an issue when leaving
the project loaded this way (assuming a manual page refresh has not
happened between steps 2 and 3).
  
Now we persist the project id when arriving at the project join / create
success pages in the onboarding.
  
### Manual page refresh causes page error

Was happening under the following sequence:
  
1. Go through onboarding, arrive in "main" app
2. Close application
3. Reopen application
4. Leave project
5. Do manual page refresh very quickly after arriving to the onboarding
page.
  
Issue was a small race condition between persisting the unset active
project id and recalculating the initial project id during the page
refresh. Persisting the active project id is an async op and when the
page reloads, we need to reinitialize the zustand store. What was
happening is that leaving the project would have the side effect of
unsetting the persisted active project id but doing the page refresh
attempts to use the process arg or what was persisted before this async
op completes, which happens to be the project that was just left if you
do it immediately. This caused a redirect to a project that you should
no longer have access to.
  
The solution is to make use of session storage that synchronizes with
the zustand store state. Session storage is persisted for the duration
of the window's lifespan (not the app), which makes it a more
appropriate mechanism for recalculating the initial active project id on
a page refresh.