fix: avoid bad project state when leaving a project fails (#377)
Fixes #376 . I made an unfortunate decision in
https://github.com/digidem/comapeo-desktop/pull/305 to let core fully
handle the checks for whether the device can leave the project. Will log
an issue in core, but basically calling `MapeoManager.leaveProject()`
has side effects despite the failure to leave a project , which causes
the bad state described in the issue. Instead, we do the relevant
membership checks client-side first and fail early if they don't pass.