fix: account for wifi ssid redaction in newer macOS versions (#380)
More of a workaround than a fix for #378 . In this case, we just display
that we can't figure out the wifi info. Think it's safe to assume that
it's unlikely that some troll uses this specific string for their wifi
SSID to spite me.

Not planning to close the issue with this PR since there's a chance that
it could be closed by some upstream fix in the future.