fix: prevent entering app for project that you are not a valid member of (#381)
Still occasionally running into situations where we end up trying to
access a project-specific page for a project that we no longer have
access to. This PR adds a route guard to make sure that we're still a
valid member of the project. If we aren't we clear the stale project ID
and reload the app at the index, which should elegantly redirect to the
appropriate location thereafter.

This doesn't feel like a fix for the root cause, but I'm also not able
to reliably reproduce the issue. Guessing there's a race condition
related to the active project ID stuff and the best we can do is to try
to get rid of potentially bad state automatically.