fix: fix accidental navigation when hovering on invite devices link in team page (#411)
There was super strange behavior occurring under the following sequence:

1. Go through invite flow
2. Navigate back to team page
3. **Hover** over `Invite Device` link

After (3), you'd end up on the devices to invite page! Guessing this is
an issue or undocumented behavior with Tanstack Router, but not entirely
sure. Seems that changing the loader for the relevant page to a
beforeLoad fixes the issue, which seems decently appropriate to do
anyways.