feat: add ability to change observation category (#413)
Towards #384

This feature is only available if you're a coordinator or the
observation of interest is one that you created.

Notable differences from design:

- The confirmation state is the green circle with a checkmark next to
the icon, instead of the alert box. This disappears 5 seconds after the
update occurs.

Known issues:

- the weird map movement after selecting a new category. not planning to
try to figure this out given upcoming changes to the map implementation
in general.