feat: update team page (#450)
Closes #424 

Notable differences from design:

- Remote archives section is pluralized and lists all active remote
archives: this is in light of significant UX flaws that the current
designs don't account for when it comes to how remote archives are
surfaced. Most of the existing designs assume that only a single remote
archive is present for a project at a time, which is incorrect and leads
to confusing UX.

- The details page for a remote archive still lists the device ID: there
isn't a strong reason to omit this as it's helpful for debugging.

Other notes:

- there's a known issue where if desktop is on this page and some
external member leaves the project (or remotes another project member
that isn't desktop), the lists may not update to reflect that. This is a
backend limitation (see
https://github.com/digidem/comapeo-core/issues/1197) and will be
addressed in a separate PR. For now, the workaround is to either
navigate in and out of the page or refresh the page.

- replaces the `chevron-right` icon with its rounded variant, which
affects the items in the comapeo settings page as well.