fix: clean up editing state when navigating away from observation (#471)
Fixes #470 

Now we clear the relevant editing state when leaving the page, whether
it's navigating to a different observation or navigating away from the
observation details altogether.

Probably need to rethink the global editing state setup at some point.
Feels pretty fragile and annoying to manage, but not sure if there are
better approaches.